requirements.txt

streamlit==1.38.0
pandas==2.2.2
matplotlib==3.8.4
python-docx==1.1.2
reportlab==4.0.9
PyPDF2==3.0.1
requests==2.32.3
beautifulsoup4==4.12.3
